+++ domaindriven-core/domaindriven-core.cabal
@@ -4,7 +4,7 @@ cabal-version: 1.12
 --
 -- see: https://github.com/sol/hpack
 
-name:           domaindriven-server
+name:           domaindriven-core
 version:        0.5.0
 synopsis:       Batteries included event sourcing and CQRS
 description:    Please see the README on GitHub at <https://github.com/tommyengstrom/domaindriven#readme>
@@ -13,8 +13,9 @@ homepage:       https://github.com/tommyengstrom/domaindriven#readme
 bug-reports:    https://github.com/tommyengstrom/domaindriven/issues
 author:         Tommy Engström
 maintainer:     tommy@tommyengstrom.com
-copyright:      2023 Tommy Engström
+copyright:      2022 Tommy Engström
 license:        BSD3
+license-file:   LICENSE
 build-type:     Simple
 extra-source-files:
     README.md
@@ -26,18 +27,14 @@ source-repository head
 
 library
   exposed-modules:
-      DomainDriven
-      DomainDriven.Internal.HasFieldName
-      DomainDriven.Internal.NamedFields
-      DomainDriven.Internal.NamedJsonFields
-      DomainDriven.Internal.Text
-      DomainDriven.Server.Class
-      DomainDriven.Server.Config
-      DomainDriven.Server.Helpers
-      DomainDriven.Server.TH
-      DomainDriven.Server.Types
+      DomainDriven.Persistance.Class
+      DomainDriven.Persistance.ForgetfulInMemory
+      DomainDriven.Persistance.Postgres
+      DomainDriven.Persistance.Postgres.Internal
+      DomainDriven.Persistance.Postgres.Migration
+      DomainDriven.Persistance.Postgres.Types
   other-modules:
-      Paths_domaindriven_server
+      Paths_domaindriven_core
   hs-source-dirs:
       src
   default-extensions:
@@ -80,19 +77,15 @@ library
     , bytestring >=0.11.3.1
     , containers >=0.6.5.1
     , deepseq >=1.4.6.1
-    , domaindriven
     , exceptions >=0.10.4
     , generic-lens >=2.2.1.0
     , http-types >=0.12.3
     , microlens >=0.4.12.0
     , mtl >=2.2.2
-    , openapi3 >=3.2.2
     , postgresql-simple >=0.6.4
     , random >=1.2.1.1
-    , servant-server >=0.19.2
     , streamly >=0.8.1.1
     , template-haskell >=2.18.0.0
-    , text >=1.2.5.0
     , time >=1.11.1.1
     , transformers >=0.5.6.2
     , unliftio >=0.2.0.1
@@ -100,4 +93,57 @@ library
     , unordered-containers >=0.2.19.1
     , uuid >=1.3.15
     , vector >=0.12.3.1
+  default-language: Haskell2010
+
+test-suite domaindriven-core-test
+  type: exitcode-stdio-1.0
+  main-is: Spec.hs
+  other-modules:
+      DomainDriven.Persistance.PostgresIORefStateSpec
+      Paths_domaindriven_core
+  hs-source-dirs:
+      test
+  default-extensions:
+      Arrows
+      ConstraintKinds
+      DataKinds
+      DeriveAnyClass
+      DeriveFunctor
+      DeriveGeneric
+      DeriveTraversable
+      DerivingStrategies
+      DuplicateRecordFields
+      FlexibleContexts
+      FunctionalDependencies
+      FlexibleInstances
+      GADTs
+      GeneralizedNewtypeDeriving
+      LambdaCase
+      MultiParamTypeClasses
+      PolyKinds
+      MultiWayIf
+      NamedFieldPuns
+      NoImplicitPrelude
+      OverloadedLabels
+      OverloadedStrings
+      RankNTypes
+      ScopedTypeVariables
+      StandaloneDeriving
+      TupleSections
+      TypeApplications
+      TypeFamilyDependencies
+      TypeOperators
+      TypeSynonymInstances
+      ViewPatterns
+  ghc-options: -Wall -Wcompat -Widentities -Wincomplete-record-updates -Wincomplete-uni-patterns -Wpartial-fields -Wredundant-constraints -Wincomplete-record-updates -Wincomplete-patterns -threaded -rtsopts -with-rtsopts=-N -Wall -Wunused-packages
+  build-depends:
+      aeson >=2.0.3.0
+    , base >=4.7 && <5
+    , domaindriven-core
+    , hspec >=2.9.7
+    , postgresql-simple >=0.6.4
+    , streamly >=0.8.1.1
+    , time >=1.11.1.1
+    , unliftio-pool >=0.2.2.0
+    , uuid >=1.3.15
   default-language: Haskell2010
