+++ domaindriven-core/package.yaml
@@ -1,10 +1,10 @@
-name:                domaindriven-server
+name:                domaindriven-core
 version:             0.5.0
 github:              "tommyengstrom/domaindriven"
 license:             BSD3
 author:              "Tommy Engström"
 maintainer:          "tommy@tommyengstrom.com"
-copyright:           "2023 Tommy Engström"
+copyright:           "2022 Tommy Engström"
 
 extra-source-files:
 - README.md
@@ -22,13 +22,6 @@ description:         Please see the README on GitHub at <https://github.com/tomm
 dependencies:
 - aeson >= 2.0.3.0
 - base >= 4.7 && < 5
-- containers >= 0.6.5.1
-- deepseq >= 1.4.6.1
-- exceptions >= 0.10.4
-- openapi3 >= 3.2.2
-- servant-server >= 0.19.2
-- text >= 1.2.5.0
-- domaindriven
 
 
 default-extensions:
@@ -86,6 +79,9 @@ library:
     - unordered-containers >= 0.2.19.1
     - vector >= 0.12.3.1
     - unliftio >= 0.2.0.1
+    - containers >= 0.6.5.1
+    - deepseq >= 1.4.6.1
+    - exceptions >= 0.10.4
 
     - streamly >= 0.8.1.1
     - postgresql-simple >= 0.6.4
@@ -96,3 +92,23 @@ library:
     - unliftio-pool >= 0.2.2.0
     - uuid >= 1.3.15
     - mtl >= 2.2.2
+tests:
+  domaindriven-core-test:
+    main:                Spec.hs
+    source-dirs:
+      - test
+    ghc-options:
+    - -threaded
+    - -rtsopts
+    - -with-rtsopts=-N
+    - -Wall
+    - -Wunused-packages
+    dependencies:
+    - domaindriven-core
+    - hspec >= 2.9.7
+    - postgresql-simple >= 0.6.4
+    - streamly >= 0.8.1.1
+    - time >=  1.11.1.1
+    - unliftio-pool >= 0.2.2.0
+    - uuid >= 1.3.15
+
