+++ domaindriven-core/src/DomainDriven/Persistance/Class.hs
