+++ domaindriven-core/src/DomainDriven/Persistance/ForgetfulInMemory.hs
