+++ domaindriven-core/test/Spec.hs
