+++ domaindriven/domaindriven.cabal
@@ -13,9 +13,8 @@ homepage:       https://github.com/tommyengstrom/domaindriven#readme
 bug-reports:    https://github.com/tommyengstrom/domaindriven/issues
 author:         Tommy Engström
 maintainer:     tommy@tommyengstrom.com
-copyright:      2022 Tommy Engström
+copyright:      2023 Tommy Engström
 license:        BSD3
-license-file:   LICENSE
 build-type:     Simple
 extra-source-files:
     README.md
@@ -27,12 +26,16 @@ source-repository head
 
 library
   exposed-modules:
-      DomainDriven.Persistance.Class
-      DomainDriven.Persistance.ForgetfulInMemory
-      DomainDriven.Persistance.Postgres
-      DomainDriven.Persistance.Postgres.Internal
-      DomainDriven.Persistance.Postgres.Migration
-      DomainDriven.Persistance.Postgres.Types
+      DomainDriven
+      DomainDriven.Internal.HasFieldName
+      DomainDriven.Internal.NamedFields
+      DomainDriven.Internal.NamedJsonFields
+      DomainDriven.Internal.Text
+      DomainDriven.Server.Class
+      DomainDriven.Server.Config
+      DomainDriven.Server.Helpers
+      DomainDriven.Server.TH
+      DomainDriven.Server.Types
   other-modules:
       Paths_domaindriven
   hs-source-dirs:
@@ -77,15 +80,19 @@ library
     , bytestring >=0.11.3.1
     , containers >=0.6.5.1
     , deepseq >=1.4.6.1
+    , domaindriven-core
     , exceptions >=0.10.4
     , generic-lens >=2.2.1.0
     , http-types >=0.12.3
     , microlens >=0.4.12.0
     , mtl >=2.2.2
+    , openapi3 >=3.2.2
     , postgresql-simple >=0.6.4
     , random >=1.2.1.1
+    , servant-server >=0.19.2
     , streamly >=0.8.1.1
     , template-haskell >=2.18.0.0
+    , text >=1.2.5.0
     , time >=1.11.1.1
     , transformers >=0.5.6.2
     , unliftio >=0.2.0.1
@@ -99,7 +106,11 @@ test-suite domaindriven-test
   type: exitcode-stdio-1.0
   main-is: Spec.hs
   other-modules:
-      DomainDriven.Persistance.PostgresIORefStateSpec
+      DomainDriven.Internal.NamedJsonFieldsSpec
+      DomainDriven.ServerSpec
+      DomainDriven.ServerSpecModel
+      Models.Counter
+      Models.Store
       Paths_domaindriven
   hs-source-dirs:
       test
@@ -137,13 +148,23 @@ test-suite domaindriven-test
       ViewPatterns
   ghc-options: -Wall -Wcompat -Widentities -Wincomplete-record-updates -Wincomplete-uni-patterns -Wpartial-fields -Wredundant-constraints -Wincomplete-record-updates -Wincomplete-patterns -threaded -rtsopts -with-rtsopts=-N -Wall -Wunused-packages
   build-depends:
-      aeson >=2.0.3.0
+      QuickCheck
+    , aeson >=2.0.3.0
+    , async
     , base >=4.7 && <5
+    , containers >=0.6.5.1
+    , deepseq >=1.4.6.1
     , domaindriven
+    , domaindriven-core
+    , exceptions >=0.10.4
     , hspec >=2.9.7
-    , postgresql-simple >=0.6.4
-    , streamly >=0.8.1.1
-    , time >=1.11.1.1
-    , unliftio-pool >=0.2.2.0
-    , uuid >=1.3.15
+    , http-client
+    , mtl
+    , openapi3 >=3.2.2
+    , quickcheck-arbitrary-adt
+    , quickcheck-classes
+    , servant-client
+    , servant-server >=0.19.2
+    , text >=1.2.5.0
+    , warp
   default-language: Haskell2010
