+++ domaindriven/package.yaml
@@ -4,7 +4,7 @@ github:              "tommyengstrom/domaindriven"
 license:             BSD3
 author:              "Tommy Engström"
 maintainer:          "tommy@tommyengstrom.com"
-copyright:           "2022 Tommy Engström"
+copyright:           "2023 Tommy Engström"
 
 extra-source-files:
 - README.md
@@ -22,6 +22,13 @@ description:         Please see the README on GitHub at <https://github.com/tomm
 dependencies:
 - aeson >= 2.0.3.0
 - base >= 4.7 && < 5
+- containers >= 0.6.5.1
+- deepseq >= 1.4.6.1
+- exceptions >= 0.10.4
+- openapi3 >= 3.2.2
+- servant-server >= 0.19.2
+- text >= 1.2.5.0
+- domaindriven-core
 
 
 default-extensions:
@@ -79,9 +86,6 @@ library:
     - unordered-containers >= 0.2.19.1
     - vector >= 0.12.3.1
     - unliftio >= 0.2.0.1
-    - containers >= 0.6.5.1
-    - deepseq >= 1.4.6.1
-    - exceptions >= 0.10.4
 
     - streamly >= 0.8.1.1
     - postgresql-simple >= 0.6.4
@@ -92,6 +96,7 @@ library:
     - unliftio-pool >= 0.2.2.0
     - uuid >= 1.3.15
     - mtl >= 2.2.2
+
 tests:
   domaindriven-test:
     main:                Spec.hs
@@ -104,11 +109,14 @@ tests:
     - -Wall
     - -Wunused-packages
     dependencies:
+    - domaindriven-core
     - domaindriven
     - hspec >= 2.9.7
-    - postgresql-simple >= 0.6.4
-    - streamly >= 0.8.1.1
-    - time >=  1.11.1.1
-    - unliftio-pool >= 0.2.2.0
-    - uuid >= 1.3.15
-
+    - QuickCheck
+    - quickcheck-arbitrary-adt
+    - quickcheck-classes
+    - warp
+    - servant-client
+    - async
+    - mtl
+    - http-client
