+++ domaindriven/src/DomainDriven/Server/Class.hs
